POC Kubernetes Python Simple Http - Docker image build and Kubernetes deploy
Creation of configuration file for Docker image build and creation of Kubernetes deployment configuration file. Creation of helper scripts to automate the build and deploy steps. The Docker image is based on a simple python image. The Kubernetes configuration is based on a Deployment specification exposed by a service specification.